% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRD4df_full}
\alias{BRD4df_full}
\title{This is data to be included in my package}
\description{
A complete HDX-MS dataset for BRD4 in apo state and in complex with IBET151
}
\author{
My Name \email{ococrook@gmail.com}
}
\keyword{data}
