% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxmsqc-utils.R
\name{chargeCorrelationHdx}
\alias{chargeCorrelationHdx}
\title{Charge states should have correlated incorperation but they need not
be exactly the same}
\usage{
chargeCorrelationHdx(object, experiment = NULL, timepoints = NULL)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{experiment}{A character vector indicating the experimental conditions}

\item{timepoints}{A numeric vector indicating the experimental timepoints}
}
\description{
Charge states should have correlated incorperation but they need not
be exactly the same
}
\author{
Oliver Crook
}
