% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxincoperationFourier.R
\name{generateSpectra}
\alias{generateSpectra}
\title{generate Spectra using a fourier transform}
\usage{
generateSpectra(
  sequences,
  incorps,
  charges,
  customs = list(code = NULL, elements = NULL)
)
}
\arguments{
\item{sequences}{A vector of peptide sequences}

\item{incorps}{A vector of deuterium incoperation}

\item{charges}{A vector of charge states of the peptide}

\item{customs}{Custom elements supplied as a list}
}
\value{
A Spectra object corresponding to the isotope distributions
}
\description{
generate Spectra using a fourier transform
}
\author{
Oliver Crook
}
