% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxincoperationFourier.R
\name{exchangeableAmides}
\alias{exchangeableAmides}
\title{Compute exchangeable amides.}
\usage{
exchangeableAmides(sequence)
}
\arguments{
\item{sequence}{The sequence of the peptide}
}
\value{
Returns a numeric indicating the number of exchangeable amides
}
\description{
Computes the number of exchangeable amides based on the sequnece
}
