\name{hdxmsqc}
\alias{hdxmsqc}
\title{A package to perfrom quality control for
  mass-spectrometry based hydrogen deuterium exchange experiment.}
\description{
  `hdxmsqc` provides the functionality to assess the quality and perform
  quality control of HDX-MS experiments. Raw and processed data can be
  visualized and analyzed to identify potential issues with the data. The 
  package is designed to work with data from any HDX-MS platform. Typically,
  users will have exported results from either HDExaminer or DynamX software.
  There is not need to filter the data in either of those software systems.
}
\author{
  Oliver Crook
}
