% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxmsqc-utils.R
\name{isMissingAtRandom}
\alias{isMissingAtRandom}
\title{Missing at random versus missing not at random}
\usage{
isMissingAtRandom(object, threshold = NULL, filter = TRUE)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{threshold}{A threshold indicated how many missing values indicate
whether missingness is not at random. Default is NULL, which means leads to a
threshold which is half the number of columns.}

\item{filter}{A logial indicating whether to filter out data that is deemed
missing not at random

data("BRD4df_full")

isMissingAtRandom(BRD4df_full)}
}
\value{
Adds a missing not at random indicator column
}
\description{
Missing at random versus missing not at random
}
\author{
Oliver Crook
}
