% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxincoperationFourier.R
\name{isotopicDistributionHDXfourier}
\alias{isotopicDistributionHDXfourier}
\title{fourier transform approach to computing isotopic distribution}
\usage{
isotopicDistributionHDXfourier(
  sequence,
  incorp = 0,
  charge = 1,
  custom = list(code = NULL, elements = NULL)
)
}
\arguments{
\item{sequence}{A peptide}

\item{incorp}{The deuterium incoperation}

\item{charge}{The charge state of the peptide}

\item{custom}{custom amino acids can be provided here provide a list
of the elements.}
}
\value{
A list of mass and intensity value corresponding to the isotope distribution
}
\description{
fourier transform approach to computing isotopic distribution
}
\examples{
isotopicDistributionHDXfourier(sequence = "HDAEHAHEAPRKL")
}
\author{
Oliver Crook
}
