% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxmsqc-utils.R
\name{plotIntensityOutliers}
\alias{plotIntensityOutliers}
\title{Intensity based deviation plot}
\usage{
plotIntensityOutliers(object, fcolIntensity = "Max.Inty")
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{fcolIntensity}{character to intensity intensity columns. Default is
"Max.Inty" and uses regular expressions to find relevant columns}
}
\value{
A ggplot2 object showing intensity based outliers
}
\description{
Intensity based deviation plot
}
\author{
Oliver Crook
}
