% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxmsqc-utils.R
\name{plotMassError}
\alias{plotMassError}
\title{Mass error plot}
\usage{
plotMassError(object, eCentroid = "Exp.Cent", tCentroid = "Theor.Cent")
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{eCentroid}{character string indicating column identifier for
experimental centroid}

\item{tCentroid}{character string indicating column identifier for
theoretical centroid}
}
\value{
a ggplot2 object which can be used to visualise the
}
\description{
Mass error plot
}
\author{
Oliver Crook
}
