% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxmsqc-utils.R
\name{plotMissing}
\alias{plotMissing}
\title{missing value plot}
\usage{
plotMissing(object, ...)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{...}{Additional arguemnts to pheatmap}
}
\value{
a pheatmap showing missing values
}
\description{
missing value plot
}
\examples{
data("BRD4df_full")
library(pheatmap)
library(RColorBrewer)

plotMissing(BRD4df_full)


}
\author{
Oliver Crook
}
