% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxmsqc-utils.R
\name{rTimeOutliers}
\alias{rTimeOutliers}
\title{Retention time based analysis}
\usage{
rTimeOutliers(
  object,
  leftRT = "leftRT",
  rightRT = "rightRT",
  searchRT = "Search.RT"
)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{leftRT}{A character indicated pattern associated with left boundary
of retention time search. Default is "leftRT".}

\item{rightRT}{A character indicated pattern associated with right boundary
of retneton time search. Default is "rightRT".}

\item{searchRT}{The actual search retention time pattern.
Default is "Search.RT"}
}
\value{
A list indicating the retention time based outliers.
}
\description{
Retention time based analysis
}
\author{
Oliver Crook
}
