% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxmsqc-utils.R
\name{computeMonotoneStats}
\alias{computeMonotoneStats}
\title{Monotonicity based outlier detection.}
\usage{
computeMonotoneStats(object, experiment = NULL, timepoints = NULL)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{experiment}{A character vector indicating the experimental conditions}

\item{timepoints}{A numeric vector indicating the experimental timepoints}
}
\description{
Monotonicity based outlier detection.
}
\author{
Oliver Crook
}
