% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxmsqc-utils.R
\name{replicateOutlier}
\alias{replicateOutlier}
\title{Correlation based checks}
\usage{
replicateOutlier(object, experiment, timepoints)
}
\arguments{
\item{experiment}{A character vector indicating the experimental conditions}

\item{timepoints}{A numeric vector indicating the experimental timepoints}
}
\value{
Returns A list of the same length as the number of experiments indicating
outlier from correlation analysis. Outliers are flagged if their deuterium
uptake is highly variable.
}
\description{
Correlation based checks
}
\author{
Oliver Crook
}
